Check erreur

+++ ft_strlen.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-int		ft_strlen(char const *str)
+size_t	ft_strlen(char const *str)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (str[i])
